tree-wide: introduce free_and_replace helper
It's a common pattern, so add a helper for it. A macro is necessary
because a function that takes a pointer to a pointer would be type specific,
similarly to cleanup functions. Seems better to use a macro.

Backport of commit 3b319885c4f to systemd-230
Only create new macro and don't apply it to existing code.

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>