fs-util: add new CHASE_SAFE flag to chase_symlinks()
When the flag is specified we won't transition to a privilege-owned
file or directory from an unprivileged-owned one. This is useful when
privileged code wants to load data from a file unprivileged users have
write access to, and validates the ownership, but want's to make sure
that no symlink games are played to read a root-owned system file
belonging to a different context.

Backport of upstream commit f14f1806e329fe92d01f15c22a384702f0cb4ae0 to systemd-230
Dropped tests

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>