macro: introduce TAKE_PTR() macro
This macro will read a pointer of any type, return it, and set the
pointer to NULL. This is useful as an explicit concept of passing
ownership of a memory area between pointers.

This takes inspiration from Rust:

https://doc.rust-lang.org/std/option/enum.Option.html#method.take

and was suggested by Alan Jenkins (@sourcejedi).

It drops ~160 lines of code from our codebase, which makes me like it.
Also, I think it clarifies passing of ownership, and thus helps
readability a bit (at least for the initiated who know the new macro)

Backport of commit ae2a15bc14bc448e625ad93fd044bc077ede4b3f to systemd-230
Just add new macro and don't switch existing code to using it.

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>