fs-util: add new CHASE_STEP flag to chase_symlinks()
If the flag is set only a single step of the normalization is executed,
and the resulting path is returned.

This allows callers to normalize piecemeal, taking into account every
single intermediary path of the normalization.

Backport of upstream commit 49eb36596b72e568d4b9b785cd4544e58c89bdee to systemd-230

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>