Refuse dbus message paths longer than BUS_PATH_SIZE_MAX limit.
Even though the dbus specification does not enforce any length limit on the
path of a dbus message, having to analyze too long strings in PID1 may be
time-consuming and it may have security impacts.

In any case, the limit is set so high that real-life applications should not
have a problem with it.

Backport of upstream commit 61397a60d98e368a5720b37e83f3169e3eb511c4 to systemd-230
CVE-2019-6454, patch 1/2

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>