RHEL: selinux: fix check for transient units
SELinux does not have a path to check for a snapshot service creation.
This ends up giving us a bogus check.

On snapshot creation we should check if the remote process type, has the
ability to start a service with the type that systemd is running with.

Based on patch from Vaclav Pavlin and Dan Walsh
http://lists.freedesktop.org/archives/systemd-devel/2013-November/014021.html

Origin: systemd-219-62.el7.src.rpm/0241-selinux-fix-check-for-transient-units.patch
Original Author: Lukas Nykryn <lnykryn@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=1296570
https://github.com/systemd/systemd/pull/1052/ (note that it was rejected)