polkit: on async pk requests, re-validate action/details
When we do an async pk request, let's store which action/details we used
for the original request, and when we are called for the second time,
let's compare. If the action/details changed, let's not allow the access
to go through.

Backport of upstream commit f697d160d6eca7ca5cc85d5391a45a4443eed177
to systemd-242 in Ubuntu 19.10 by Ubuntu devs

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>