polkit: when authorizing via PK let's re-resolve callback/userdata instead of caching it
Previously, when doing an async PK query we'd store the original
callback/userdata pair and call it again after the PK request is
complete. This is problematic, since PK queries might be slow and in the
meantime the userdata might be released and re-acquired. Let's avoid
this by always traversing through the message handlers so that we always
re-resolve the callback and userdata pair and thus can be sure it's
up-to-date and properly valid.

Backport of upstream commit 2ebee744d9e1c4a28e82a945f94e8309c0b3543f
to systemd-242 in Ubuntu 19.10 by Ubuntu devs

Signed-off-by: Mikhail Novosyolov <m.novosyolov@rosalinux.ru>