Drop warning about depreceated /var/run
Unfortunately there is no easy, good and reliable way to change /var/run to /run
in all packages. As a result, too many messages about this are printed and may
confuse users or just take too much space on the screen and make other more
important messages not seeable.

There is an option to add such a replacement to spec-helper, but it is a
not very reliable way to change things and will allow to leave /var/run in even
self-written systemd units and configs which are attached to a package.